This is why i hate modding with fabric and forge at the same time

+++ common/src/main/java/de/rafael/mods/chronon/technology/interfaces/ChrononStorageItem.java
@@ -9,9 +9,9 @@ import net.minecraft.world.item.ItemStack;
 
 public interface ChrononStorageItem {
 
-    long chronons(ItemStack itemStack);
-    void chronons(ItemStack itemStack, long amount);
+    int chronons(ItemStack itemStack);
+    void chronons(ItemStack itemStack, int amount);
 
-    void removeChronons(ItemStack itemStack, long amount);
+    void removeChronons(ItemStack itemStack, int amount);
 
 }
